parse function refactored and passing norm now.

+++ include/wrapper.h
@@ -6,7 +6,7 @@
 /*   By: mogawa <mogawa@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/19 15:56:18 by mogawa            #+#    #+#             */
-/*   Updated: 2024/02/20 09:31:54 by mogawa           ###   ########.fr       */
+/*   Updated: 2024/02/29 13:44:16 by mogawa           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 # include <stdlib.h>
 
 void	ft_perror_exit(int status, char const *msg);
-double	ft_xatod(char const * const str_num);
-double	ft_ranged_xatod(char const * const str_num, int min, int max);
-char**	ft_xsplit(char const *s, char c, size_t num_childs);
+double	ft_xatod(char const *const str_num);
+double	ft_ranged_xatod(char const *const str_num, double min, double max);
+char	**ft_xsplit(char const *s, char c, size_t num_childs);
 
 #endif
